feat: move allTweetsDummyData to AuthContext
